Refactor strip_source_roots.py
* Rename file to strip_source_roots.py
* Rename rule to emphasize that this comes from a target
* Early return when the target type is Files
* Move some variables around
* Rename some test methods
