+++ src/python/pants/backend/python/rules/run_setup_py_test.py
@@ -39,7 +39,7 @@ from pants.engine.rules import RootRule
 from pants.engine.scheduler import ExecutionError
 from pants.engine.selectors import Params
 from pants.python.python_requirement import PythonRequirement
-from pants.rules.core.strip_source_root import strip_source_root
+from pants.rules.core.strip_source_roots import strip_source_roots_from_target
 from pants.source.source_root import SourceRootConfig
 from pants.testutil.subsystem.util import init_subsystem
 from pants.testutil.test_base import TestBase
@@ -68,7 +68,7 @@ class TestGenerateChroot(TestSetupPyBase):
       generate_chroot,
       get_sources,
       get_requirements,
-      strip_source_root,
+      strip_source_roots_from_target,
       get_ancestor_init_py,
       get_owned_dependencies,
       get_exporting_owner,
@@ -158,7 +158,7 @@ class TestGetSources(TestSetupPyBase):
   def rules(cls):
     return super().rules() + [
       get_sources,
-      strip_source_root,
+      strip_source_roots_from_target,
       get_ancestor_init_py,
       RootRule(SetupPySourcesRequest),
       RootRule(SourceRootConfig),
