+++ src/python/pants/backend/python/rules/setup_py_util.py
@@ -12,7 +12,7 @@ from pants.engine.fs import FilesContent
 from pants.engine.legacy.graph import HydratedTarget
 from pants.engine.legacy.structs import PythonTargetAdaptor, ResourcesAdaptor
 from pants.python.python_setup import PythonSetup
-from pants.rules.core.strip_source_root import SourceRootStrippedSources
+from pants.rules.core.strip_source_roots import SourceRootStrippedSources
 from pants.source.source_root import SourceRoots
 from pants.util.strutil import ensure_text
