+++ src/python/pants/rules/core/strip_source_roots_test.py
@@ -9,8 +9,8 @@ from pants.build_graph.files import Files
 from pants.engine.legacy.graph import HydratedTarget
 from pants.engine.rules import RootRule
 from pants.engine.selectors import Params
-from pants.rules.core.strip_source_root import SourceRootStrippedSources
-from pants.rules.core.strip_source_root import rules as strip_source_root_rules
+from pants.rules.core.strip_source_roots import SourceRootStrippedSources
+from pants.rules.core.strip_source_roots import rules as strip_source_root_rules
 from pants.testutil.option.util import create_options_bootstrapper
 from pants.testutil.test_base import TestBase
 
@@ -20,7 +20,7 @@ class StripSourceRootsTests(TestBase):
   def rules(cls):
     return (*super().rules(), *strip_source_root_rules(), RootRule(HydratedTarget))
 
-  def assert_stripped_source_file(
+  def assert_stripped_target(
     self, *, original_path: str, expected_path: str, target_type_alias: Optional[str] = None,
   ) -> None:
     adaptor = Mock()
@@ -36,19 +36,19 @@ class StripSourceRootsTests(TestBase):
     )
     self.assertEqual(stripped_sources.snapshot.files, (expected_path,))
 
-  def test_source_roots_python(self):
-    self.assert_stripped_source_file(
+  def test_strip_python_target(self):
+    self.assert_stripped_target(
       original_path='src/python/pants/util/strutil.py', expected_path='pants/util/strutil.py',
     )
 
-  def test_source_roots_java(self):
-    self.assert_stripped_source_file(
+  def test_strip_java_target(self):
+    self.assert_stripped_target(
       original_path='src/java/some/path/to/something.java',
       expected_path='some/path/to/something.java',
     )
 
-  def test_dont_strip_source_for_files(self):
-    self.assert_stripped_source_file(
+  def test_dont_strip_files_target(self):
+    self.assert_stripped_target(
         original_path='src/python/pants/util/strutil.py',
         expected_path='src/python/pants/util/strutil.py',
         target_type_alias=Files.alias()
